feat: add support for chunking server worker results to increase max supported mapping size

+++ src/spawn-worker.js
@@ -3,6 +3,7 @@ import Worker from 'worker-loader!./suggestion.worker.js' // eslint-disable-line
 export default function spawnSuggestionWorker(source, target, mappingType) {
   return new Promise((resolve) => {
     const worker = new Worker()
+    let result = {}
 
     worker.postMessage({
       type: 'suggestions',
@@ -10,9 +11,13 @@ export default function spawnSuggestionWorker(source, target, mappingType) {
     })
 
     worker.onmessage = ({ data }) => {
+      if (data.slice(0, 8) === 'COMPLETE') {
+        console.log('Recieved data chunk:', data)
+        worker.terminate()
+        return resolve(result)
+      }
       const parsed = JSON.parse(data)
-      worker.terminate()
-      return resolve(parsed)
+      result = {...result, ...parsed}
     }
   })
 }
